refactor: move `Level` to `logging` subpackage
