+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/core/logging/Level.kt
@@ -1,0 +1,10 @@
+package org.firstinspires.ftc.teamcode.core.logging
+
+import android.util.Log
+
+enum class Level(val priority: Int) {
+    Debug(Log.DEBUG),
+    Info(Log.INFO),
+    Warn(Log.WARN),
+    Error(Log.ERROR),
+}
